\name{binaryar}
\alias{binaryar}
\title{Multivariate Binary Test Data Set 2}
\usage{data(binaryar)}
\description{
  A simulated data set with a binary response y, a covariate x, and a cluster 
  group variable id. Used in examples for mprobit function.
  The values used for the simulation are:
  beta0=-0.3, beta1=0.4, latent AR(1) correlation = 0.7
}

\keyword{datasets}


