\name{binaryex}
\alias{binaryex}
\title{Multivariate Binary Test Data Set 1}
\usage{data(binaryex)}
\description{
  A simulated data set with a binary response y, a covariate x, and a cluster 
  group variable id. Used in examples for mprobit function.
  The values used for the simulation are:
  beta0=0.3, beta1=0.2, latent exchangeable correlation = 0.6
}

\keyword{datasets}


