\name{ordinalex}
\alias{ordinalex}
\title{Multivariate Ordinal Test Data Set}
\usage{data(ordinalex)}
\description{
  A simulated data set with an ordinal response y (3 categories), 
  a covariate x, and a cluster group variable id. 
  Used in examples for ordprobit function.
  The values used for the simulation are:
  cutpt1=-0.5, cutpt2=0.3, beta1=0.4, latent exchangeable correlation = 0.6
}

\keyword{datasets}


